refactor(project): move init and update logic into project classes
