+++ bin/piqule.php
@@ -2,17 +2,18 @@
 
 declare(strict_types=1);
 
-use Haspadar\Piqule\Init;
 use Haspadar\Piqule\Output\Console;
 use Haspadar\Piqule\Output\Line\Error;
 use Haspadar\Piqule\PiquleException;
-use Haspadar\Piqule\PiquleSentinel;
+use Haspadar\Piqule\Project\InitializedProject;
+use Haspadar\Piqule\Project\ProjectOf;
+use Haspadar\Piqule\Project\Sentinel;
+use Haspadar\Piqule\Project\UninitializedProject;
 use Haspadar\Piqule\RunContext;
 use Haspadar\Piqule\Source\DiskSourceDirectory;
 use Haspadar\Piqule\Step\End;
 use Haspadar\Piqule\Step\MissingTarget;
 use Haspadar\Piqule\Target\DiskTargetDirectory;
-use Haspadar\Piqule\Update;
 
 require_once dirname(__DIR__) . '/vendor/autoload.php';
 
@@ -20,24 +21,24 @@ $output = new Console();
 
 try {
     $context = new RunContext($argv);
-
-    $templates = dirname(__DIR__) . '/templates';
-    $root = $context->root();
-
-    $sentinel = new PiquleSentinel($context->root());
-
-    match ($context->argument(1)) {
-        'init' => (new Init(
-            new DiskSourceDirectory($templates),
-            new DiskTargetDirectory($root),
+    $sourceDirectory = new DiskSourceDirectory(dirname(__DIR__) . '/templates');
+    $targetDirectory = new DiskTargetDirectory($context->root());
+    $project = new ProjectOf(
+        new Sentinel($context->root()),
+        new InitializedProject($output),
+        new UninitializedProject(
+            $sourceDirectory,
+            $targetDirectory,
             new MissingTarget(
                 $output,
                 new End($output),
             ),
-        ))->run(),
+        ),
+    );
 
-        'update' => (new Update($output))->run(),
-
+    match ($context->argument(1)) {
+        'init' => $project->init(),
+        'update' => $project->update(),
         default => throw new PiquleException(
             sprintf('Unknown command: %s', $context->argument(1)),
         ),
